refactor(examples): consolidate static-website and custom-domain-website
Merges the two static-website examples into a single canonical one. The
old static-website demonstrated S3 origin + OAC + bucket deployment + SNS
alerting; the old custom-domain-website demonstrated Route 53 + ACM + apex
alias records. Both produced a static website on CloudFront but each
omitted half the story. The combined example composes all of it in one
graph — zone → cert → S3 site → cdn → alias records + deploy — so every
cross-component dependency is a ref() and the dependency resolver enforces
build order. Removes the custom-domain-website source, test, and
bin/app.ts entry and updates the static-website snapshot.
